+++ src/Controller/UserRestController.php
@@ -64,6 +64,15 @@ class UserRestController extends AbstractFOSRestController
 	}
 
 	/**
+	 * @Rest\Post("user/logout")
+	 */
+	public function logoutAction(ParamFetcher $paramFetcher): Response
+	{
+		$this->session->logout();
+		return $this->handleView($this->view([], 200));
+	}
+
+	/**
 	 * @Rest\Delete("user/{userId}", requirements={"userId" = "\d+"})
 	 */
 	public function deleteUserAction(int $userId): Response