+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - Ambassadors of Austria are automatically included to the Austrian AMB working group @peter.toennies
 - Adjust devdocs to being open source !823 @flukx
 - Mention test artifacts under „Troubleshooting in devdocs“ !845 @flukx
+- Added logout Rest endpoint !866 @alex.simm
 
 ## Bugfixes
 - Orga can delete quizzes #364 !767 @k.miklobusec